update: auth token

+++ src/main/java/com/example/demo/service/UserService.java
@@ -1,11 +1,15 @@
 package com.example.demo.service;
 
 import com.example.demo.dto.request.UserCreationRequest;
+import com.example.demo.dto.response.UserResponse;
 import com.example.demo.entity.User;
 import com.example.demo.exception.AppException;
 import com.example.demo.exception.ErrorCode;
+import com.example.demo.mapper.UserMapper;
 import com.example.demo.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,29 +18,39 @@ import java.util.List;
 public class UserService {
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private UserMapper userMapper;
+
     public User createUser(UserCreationRequest request) {
-        User user = new User();
-        if(userRepository.existsByName(request.getName())) {
+
+        if (userRepository.existsByName(request.getName())) {
             throw new AppException(ErrorCode.USER_EXITED);
         }
         if (userRepository.existsByEmail(request.getEmail())) {
             throw new AppException(ErrorCode.EMAIL_EXITED);
         }
-        user.setName(request.getName());
-        user.setPassword(request.getPassword());
-        user.setEmail(request.getEmail());
-        user.setDob(request.getDob());
+        User user = userMapper.toUser(request);
+        PasswordEncoder passwordEncoder = new BCryptPasswordEncoder(10);
+        user.setPassword(passwordEncoder.encode(request.getPassword()));
         return userRepository.save(user);
     }
+
     public List<User> getAllUsers() {
         return userRepository.findAll();
     }
-    public User getUserById(String id) {
-        return userRepository.findById(id).orElseThrow(() -> new RuntimeException("Không tìm thấy người dùng với id: " + id));
+
+    public UserResponse getUserById(String id) {
+        return userMapper.toUserResponse(userRepository.findById(id).orElseThrow(() -> new RuntimeException("Không tìm thấy người dùng với id: " + id)));
     }
+
     public void deleteUserById(String id) {
-         userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("Không tìm thấy người dùng với id: " + id);
+        }
+        userRepository.deleteById(id);
     }
+
     public User updateUser(String id, UserCreationRequest request) {
         User user = userRepository.findById(id).orElse(null);
         if (user != null) {
